run tests twice, check for determinism

+++ pkg/bramble/integration_test.go
@@ -7,6 +7,7 @@ import (
 	"testing"
 
 	"github.com/alecthomas/assert"
+	"github.com/maxmcd/bramble/pkg/reptar"
 )
 
 var (
@@ -14,18 +15,37 @@ var (
 )
 
 func TestIntegration(t *testing.T) {
-	b := Bramble{}
-
+	runTests := func() {
+		b := Bramble{}
+		if err := b.test([]string{"../../tests"}); err != nil {
+			fmt.Printf("%+v", err)
+			t.Error(err)
+		}
+	}
+	hasher := NewHasher()
 	dir, err := ioutil.TempDir("", TestTmpDirPrefix)
+	assert.NoError(t, err)
+	hasher2 := NewHasher()
+	dir2, err := ioutil.TempDir("", TestTmpDirPrefix)
 	assert.NoError(t, err)
 	// set a unique bramble store for these tests
 	os.Setenv("BRAMBLE_PATH", dir)
-	if err := b.test([]string{"../../tests"}); err != nil {
-		fmt.Printf("%+v", err)
+	runTests()
+	if err = reptar.Reptar(dir+"/store", hasher); err != nil {
 		t.Error(err)
+	}
+	os.Setenv("BRAMBLE_PATH", dir2)
+	runTests()
+	if err = reptar.Reptar(dir2+"/store", hasher2); err != nil {
+		t.Error(err)
+	}
+	if hasher.String() != hasher2.String() {
+		t.Error("content doesn't match, non deterministic", dir, dir2)
+		return
 	}
 
 	_ = os.RemoveAll(dir)
+	_ = os.RemoveAll(dir2)
 }
 
 func TestRunStarlarkBuilder(t *testing.T) {
